[add] music page
With tone.js there is now a hook for a sampler object that can
be used to play sounds.

I tried to overhaul a lot of the component type inheritance but
it is a huge mess and I will have to fix it later.

TO FIX:
type inheritance, overlapping/redundant component
functionalities, redundant CSS, style cleanup, probably
other things I'm forgetting

Also need to put in bars dividing the music sequencer into
groups of 4 keys.